+++ visitz/VisitzApi/Endpoints/Caseload/GetCaseloadEndpoint.cs
@@ -4,7 +4,7 @@ using VisitzApi.Models.Caseload;
 
 #nullable enable
 
-namespace VisitzApi.Endpoints;
+namespace VisitzApi.Endpoints.Caseload;
 
 internal class GetCaseloadEndpoint(string baseUrl, DateTimeOffset? after = null)
     : VisitzBaseEndpoint<CaseloadJson>(baseUrl, Vpi.V2, CaseloadPath)
