+++ visitz/VisitzApi/Vpi.cs
@@ -1,5 +1,6 @@
 using VisitzApi.Endpoints;
 using VisitzApi.Endpoints.Attachments;
+using VisitzApi.Endpoints.Caseload;
 using VisitzApi.Endpoints.Notes;
 using VisitzApi.Endpoints.SafetyAssess;
 using VisitzApi.Endpoints.Visits;
